requirements.txt
absl-py==0.7.1
astor==0.8.0
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
blis==0.2.4
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cycler==0.10.0
cymem==2.0.2
Cython==0.29.9
decorator==4.4.0
defusedxml==0.6.0
en-core-web-md==2.1.0
en-core-web-sm==2.1.0
entrypoints==0.3
Flask==1.0.3
Flask-CacheBuster==1.0.0
gast==0.2.2
grpcio==1.21.1
h5py==2.9.0
idna==2.8
imageio==2.5.0
imutils==0.5.2
ipykernel==5.1.1
ipython==7.5.0
ipython-genutils==0.2.0
ipywidgets==7.4.2
isort==4.3.20
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10.1
joblib==0.13.2
jsonschema==3.0.1
jupyter==1.0.0
jupyter-client==5.2.4
jupyter-console==6.0.0
jupyter-core==4.4.0
Keras==2.0.3
Keras-Applications==1.0.7
Keras-Preprocessing==1.0.9
kiwisolver==1.1.0
lazy-object-proxy==1.4.1
lxml==4.3.3
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.0
mccabe==0.6.1
mistune==0.8.4
mock==3.0.5
murmurhash==1.0.2
nbconvert==5.5.0
nbformat==4.4.0
networkx==2.3
nltk==3.4.1
notebook==5.7.8
numpy==1.16.3
openalpr==1.0.12
opencv-python==4.1.0.25
pandas==0.24.2
pandocfilters==1.4.2
parso==0.4.0
pdf2image==1.5.4
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.0.0
plac==0.9.6
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.8.0
ptyprocess==0.6.0
Pygments==2.4.0
pylint==2.3.1
pyparsing==2.4.0
PyQt5==5.12.2
PyQt5-sip==4.19.17
pyrsistent==0.15.2
pytesseract==0.2.6
python-dateutil==2.8.0
pytz==2019.1
PyWavelets==1.0.3
PyYAML==5.1
pyzmq==18.0.1
qtconsole==4.4.4
requests==2.22.0
resources==0.0.1
scikit-image==0.15.0
scikit-learn==0.21.2
scipy==1.3.0
Send2Trash==1.5.0
simplejson==3.16.0
six==1.12.0
sklearn==0.0
spacy==2.1.4
srsly==0.0.5
staty==1.2.3
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-estimator==1.13.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
Theano==1.0.4
thinc==7.0.4
tornado==6.0.2
tqdm==4.32.1
traitlets==4.3.2
typed-ast==1.4.0
urllib3==1.25.3
wasabi==0.2.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
widgetsnbextension==3.4.2
wrapt==1.11.1
